% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Areal_level_data.R
\docType{data}
\name{Areal_level_data}
\alias{Areal_level_data}
\title{Areal level data}
\format{
A data frame with 25 rows and 2 columns:
\describe{
\item{Region}{Region number}
\item{Pop}{Population of each region}
\item{Ind}{percentage of indigenous people in each region}
...
}
}
\usage{
Areal_level_data
}
\description{
The data which describes the areal level data (proportion of indigenous people and population) for 25 regions.
}
\keyword{datasets}
