% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_level_data.R
\docType{data}
\name{Individual_level_data}
\alias{Individual_level_data}
\title{Individual level data}
\format{
A data frame with 758 rows and 4 columns:
\describe{
\item{Index}{Index for individual}
\item{Ave_Age_5}{Average of age smaller than 5}
\item{Ave_Age_65}{Average of age higher 65}
\item{SEFI}{SEFI factor}
...
}
}
\usage{
Individual_level_data
}
\description{
The data which describes the Individual level data (average age below 5, average age above 65 and SEFI factor) for 758 individuals.
}
\keyword{datasets}
