% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regional_data.R
\docType{data}
\name{Regional_data}
\alias{Regional_data}
\title{Regional data}
\format{
A data frame with 758 rows and 7 columns:
\describe{
\item{Index}{Index for individual}
\item{Lat}{Latitude of the area}
\item{Long}{Longitude of the area}
\item{Population}{Population of each area}
\item{SEFI}{SEFI factor}
\item{Age_5}{Number of individuals age below 5}
\item{Age_65}{Number of individuals age above 65}
...
}
}
\usage{
Regional_data
}
\description{
The data which describes the individual level data and areal data for 758 individuals.
}
\keyword{datasets}
